Python FastAPI Backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
torch>=2.0.0
transformers>=4.35.0
soundfile>=0.12.0
kokoro-tts>=1.0.0
requests>=2.31.0